+++ src/components/layout/footer.tsx
@@ -1,0 +1,3 @@
+export const Footer = () => {
+  return <footer className={'w-full p-2 bg-gray-300'}></footer>;
+};