+++ src/pages/_app.tsx
@@ -1,9 +1,14 @@
-import '../styles/globals.css';
+import { Layout } from '@/components/layout';
+import '@/styles/globals.css';
 import 'antd/dist/antd.css';
 import type { AppProps } from 'next/app';
 
 function MyApp({ Component, pageProps }: AppProps) {
-  return <Component {...pageProps} />;
+  return (
+    <Layout>
+      <Component {...pageProps} />
+    </Layout>
+  );
 }
 
 export default MyApp;